weights/requirements.txt
streamlit
numpy 
pytorch
uvicorn
fastapi
pydantic
torchvision
transformers
